Bump dependencies as advised by NPM audit
When I added the SDK to my project, NPM warned me of security
vulnerabilites in some dependencies in the SDK.

This PR updates the dependencies by running the command `npm audit fix`.

I note, however, PRs #62, #61, #60, and #59 have been opened, which may
cause merge conflicts if they are merged as well as this commit. It
would be excellent to merge these PRs too if possible, as they solve
potential security holes.

Signed-off-by: Dom Rodriguez <shymega@shymega.org.uk>